requirements.txt
requests
pandas
scikit-learn
matplotlib
python-dotenv